created how to play overlay
