+++ assets/css/styles.css
@@ -45,21 +45,24 @@ h1 {
   width: 100%;
   height: 100%;
   top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background-color: rgba(0, 0, 0, 0.5);
+  background-color: rgba(0, 0, 0, 0.8);
   z-index: 2;
   cursor: pointer;
 }
 .overlay-text {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  font-size: 50px;
+  padding: 10%;
   color: white;
-  transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
+}
+.overlay-text > h2 {
+  margin: 3% auto;
+  text-align: center;
+}
+.overlay-text > ul {
+  list-style: bullet;
+  padding: 2%;
+}
+.overlay-text li {
+  display: list-item;
 }
 
 /*hangman image and keydown letter*/
@@ -162,7 +165,7 @@ footer {
   background-color: #b7bdc4;
   width: 100vw;
 }
-footer ul {
+footer > ul {
   display: flex;
   list-style-type: none;
   justify-content: space-between;
